services/embedding-bot/requirements.txt
fastapi
uvicorn
httpx
pydantic
google-genai
chromadb